requirements.txt
adafruit-circuitpython-typing
pytest
